requirements.txt
requests
apscheduler
python-dotenv
openai
